+++ branches/SFML2/src/game_type_select.cpp
@@ -121,7 +121,7 @@ void CGameTypeSelect::Enter() {
 	Music.Play(param.menu_music, true);
 }
 
-void CGameTypeSelect::Loop(double time_step) {
+void CGameTypeSelect::Loop(float time_step) {
 	ScopedRenderMode rm(GUI);
 	Winsys.clear();
