+++ branches/SFML2/src/loading.h
@@ -23,7 +23,7 @@ GNU General Public License for more details.
 
 class CLoading : public State {
 	void Enter();
-	void Loop(double time_step);
+	void Loop(float time_step);
 public:
 };
