+++ branches/SFML2/src/splash_screen.h
@@ -23,7 +23,7 @@ GNU General Public License for more details.
 
 class CSplashScreen : public State {
 	void Enter();
-	void Loop(double time_step);
+	void Loop(float time_step);
 public:
 };
